# Optimized FreeList allocator