--- a/tt_metal/impl/allocator/algorithms/free_list_opt.cpp
+++ b/tt_metal/impl/allocator/algorithms/free_list_opt.cpp
@@ -134,10 +134,11 @@ std::optional<DeviceAddr> FreeListOpt::allocate_at_address(DeviceAddr absolute_s
     // Nothing we can do but scan the free list
     size_t alloc_size = align(std::max(size_bytes, min_allocation_size_));
     ssize_t target_block_index = -1;
+    DeviceAddr start_address = absolute_start_address - offset_bytes_;
     for (size_t i = 0; i < block_address_.size(); i++) {
         size_t block_start = block_address_[i];
         size_t block_end = block_start + block_size_[i];
-        if (absolute_start_address >= block_start && absolute_start_address + alloc_size <= block_end) {
+        if (start_address >= block_start && start_address + alloc_size <= block_end) {
             target_block_index = i;
             break;
         }
@@ -154,9 +155,9 @@ std::optional<DeviceAddr> FreeListOpt::allocate_at_address(DeviceAddr absolute_s
     TT_ASSERT(it != segregated_list.end(), "Block not found in size segregated list");
     segregated_list.erase(it);
 
-    size_t offset = absolute_start_address - block_address_[target_block_index];
+    size_t offset = start_address - block_address_[target_block_index];
     size_t alloc_block_index = allocate_in_block(target_block_index, alloc_size, offset);
-    return block_address_[alloc_block_index] + offset_bytes_;
+    return absolute_start_address;
 }
 
 size_t FreeListOpt::allocate_in_block(size_t block_index, DeviceAddr alloc_size, size_t offset) {