--- a/tt_metal/impl/allocator/allocator.cpp
+++ b/tt_metal/impl/allocator/allocator.cpp
@@ -7,7 +7,7 @@
 #include <magic_enum/magic_enum.hpp>
 #include "tt_metal/common/math.hpp"
 #include "tt_metal/detail/util.hpp"
-#include "tt_metal/impl/allocator/algorithms/free_list.hpp"
+#include "tt_metal/impl/allocator/algorithms/free_list_opt.hpp"
 #include "tt_metal/impl/buffers/buffer.hpp"
 
 namespace tt {
@@ -28,7 +28,7 @@ static char const* get_memory_pool_name(BufferType buffer_type) {
 
 void BankManager::init_allocator(DeviceAddr size_bytes, uint32_t alignment_bytes, DeviceAddr offset) {
     this->allocator_ =
-        std::make_unique<FreeList>(size_bytes, offset, alignment_bytes, alignment_bytes, FreeList::SearchPolicy::FIRST);
+        std::make_unique<FreeListOpt>(size_bytes, offset, alignment_bytes, alignment_bytes);
 }
 
 void validate_num_banks(uint32_t num_banks, const BufferType& buffer_type, bool disable_interleaved) {