@abhullar-tt 2024-11-28
Hey Marty, this is awesome! The unoptimized allocator was something on the back of my mind for a while. Thanks for taking this on ⭐