@abhullar-tt 2024-11-28
Running tests on CI [here](https://github.com/tenstorrent/tt-metal/actions?query=branch%3Aabhullar%2F15536) 

Would it be possible to benchmark fragmentation @marty1885? I don't think we would be worse off since the current allocator is set to use first fit.