@marty1885 2024-11-28
@abhullar-tt Let me debug that. It shouldn't though since it's in best mode
 instead of first.