@marty1885 2024-11-29
@abhullar-tt How can I see the exact error message and cause? Like know when the allocator reported OOM? I can't find indications on the CI. I will only have access to my QB during weekend, have to wait for that. Some networking issues where I host it.

There is a known pattern causing BestFit to have more fragmentation then FirstFit. If a fixed allocation is wedged in the middle-to-side of the allocatable space, the BestFit allocator may see one half of the remaining space as smaller and dump everything there. Since technically it is a better fit.