@abhullar-tt 2024-12-02
> @abhullar-tt can you run the CI again? I have confirmed the UT is passing and LLMs are running on my machine. I am not 100% all tests passes as I lack N150 locally.

Starting running some [tests](https://github.com/tenstorrent/tt-metal/actions?query=branch%3Aallocator-opt)
Do you think you could add some benchmarks on fragmentation differences between the two allocator algos?