@marty1885 2024-12-03
> Do you think you could add some benchmarks on fragmentation differences between the two allocator algos?

The earlier problem stems from a bug not discovered in testing (deallocate() not working when offset !=0). Otherwise the new allocator is equivalent to the existing FreeList with the BestFit policy. I made it this way to reduce the complication of figuring out if fragmentation increased under TT's use case (which the allocation pattern is not typical).

I can add some fragmentation benchmark if it is required. What do you imagine it be like?