@marty1885 2024-12-04
NP. I wrote a quick test comparing my new allocator against the existing one. To simulate random allocation and deallocation on WH's L1. Random allocations are performed and there's a chance of deallocation of previous allocations. Measuring how many iterations has to happen until the memory is full.

Results are as shown. My allocator is indded the same as the existing FreeList with BestFit policy (and consistently better then FirstFit used now). Actually I found a bug in the existing BestFit that is sometimes messes up and doesn't pick the best block. So rarely my allocator will perform slightly better is rare cases

```
Benchmarking fragmentation... (number of random allocation and deallocation attempts until full)
FreeListOpt: 337
FreeList (First): 321
FreeList (Best): 337
```

Source code can be found here: https://github.com/marty1885/tt-alloc-opt/blob/master/fragmentation.cpp