@abhullar-tt 2024-12-06
> @abhullar-tt Ping on merging this! I see some exciting features merged recently into main and wants to take advantage of them. But also I need the faster allocator.

Yep, this is on top of my list! I held off earlier this week because we had some instability in our CI.