@abhullar-tt 2024-12-06
@marty1885 do you think you could take a look at the tests failing in: https://github.com/tenstorrent/tt-metal/actions/runs/12202495523
Earlier I thought it was machine instability but on multiple attempts it looks like these tests crash with the changes here.