@abhullar-tt 2024-12-09
> @abhullar-tt Thanks for finding the problem. Should be fixed with the latest commit. Please run the CI again.

Getting closer but still some failures: https://github.com/tenstorrent/tt-metal/actions/runs/12227847709. 
I think you can try to run this locally by invoking the same tests as `all-post-commit-workflows.yaml`